[FEAT]: Added store function for department + route store

+++ api-backend/database/migrations/2023_08_01_142631_create_employee_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('lastName', 100)->nullable(false);
             $table->string('email', 60)->nullable(false)->unique();
             $table->string('phone', 20)->nullable(false);
-            $table->integer('department_id')->nullable(false);
+            $table->unsignedBigInteger('department_id')->nullable(false);
 
         });
